third_party/blink/web_tests/platform/mac/external/wpt/html/canvas/element/text/2d.text.measure.fontBoundingBox-reduced-ascent-expected.txt

This is a testharness.js-based test.
[FAIL] Testing fontBoundingBox for OffscreenCanvas with reduced ascent metric
  assert_equals: ctx.measureText('A').fontBoundingBoxAscent === 10 (got 68[number], expected 10[number]) expected 10 but got 68
Harness: the test ran to completion.
